Swift tools 6.1 and new file structure
