+++ Sources/SwiftUIBackports/iOS18/Backport+ImagePlayground.swift
@@ -1,0 +1,30 @@
+//
+//  Backport+ImagePlayground.swift
+//  SwiftUIBackports
+//
+//  Created by Ricky Witherspoon on 7/2/25.
+//
+
+import SwiftUI
+import ImagePlayground
+
+@MainActor
+public extension Backport where Content: View {
+    @ViewBuilder func imagePlayground(
+        _ presented: Binding<Bool>,
+        completion: @escaping (URL?) -> Void
+    ) -> some View {
+        if #available(iOS 18.1, *) {
+            if ImagePlaygroundViewController.isAvailable {
+                content
+                    .imagePlaygroundSheet(isPresented: presented) { url in
+                        completion(url)
+                    }
+            } else {
+                content
+            }
+        } else {
+            content
+        }
+    }
+}
